+++ gsa/src/web/graphql/__tests__/wizards.js
@@ -25,9 +25,15 @@ import {GraphQLError} from 'graphql';
 
 import date, {setLocale} from 'gmp/models/date';
 
+import {hasValue} from 'gmp/utils/identity';
+
 import {rendererWith, screen, wait, fireEvent} from 'web/utils/testing';
 
-import {useRunQuickFirstScan, useRunModifyTask} from '../wizards';
+import {
+  useRunQuickFirstScan,
+  useRunModifyTask,
+  useRunQuickTask,
+} from '../wizards';
 
 import {
   createWizardTargetQueryMock,
@@ -36,6 +42,8 @@ import {
   createWizardScheduleQueryMock,
   createWizardAlertQueryMock,
   createWizardModifyTaskQueryMock,
+  createAdvancedWizardCreateTaskQueryMock,
+  createAdvancedWizardTargetQueryMock,
 } from '../__mocks__/wizards';
 
 setLocale('en'); // Required for composing wizard entity name
@@ -187,11 +195,13 @@ const RunModifyTaskComponent = ({alertEmail, reschedule}) => {
 describe('useRunModifyTask tests', () => {
   test('Should create schedule, alert, and modify task after user interaction', async () => {
     const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'myFirstTask',
       mockCreationDate,
       mockScheduleDate,
       mockScheduleTimezone,
     );
     const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'myFirstTask',
       mockCreationDate,
     );
     const [modifyTaskMock, modifyTaskResult] = createWizardModifyTaskQueryMock(
@@ -233,12 +243,14 @@ describe('useRunModifyTask tests', () => {
       scheduleMock,
       scheduleResult,
     ] = createWizardScheduleQueryMock(
+      'myFirstTask',
       mockCreationDate,
       mockScheduleDate,
       mockScheduleTimezone,
       [error],
     );
     const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'myFirstTask',
       mockCreationDate,
     );
     const [modifyTaskMock, modifyTaskResult] = createWizardModifyTaskQueryMock(
@@ -280,11 +292,13 @@ describe('useRunModifyTask tests', () => {
 
   test('Should not create a schedule if reschedule is 0', async () => {
     const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'myFirstTask',
       mockCreationDate,
       mockScheduleDate,
       mockScheduleTimezone,
     );
     const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'myFirstTask',
       mockCreationDate,
     );
     const [modifyTaskMock, modifyTaskResult] = createWizardModifyTaskQueryMock(
@@ -322,11 +336,13 @@ describe('useRunModifyTask tests', () => {
 
   test('Should not create an alert if alert_email is empty string', async () => {
     const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'myFirstTask',
       mockCreationDate,
       mockScheduleDate,
       mockScheduleTimezone,
     );
     const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'myFirstTask',
       mockCreationDate,
     );
     const [modifyTaskMock, modifyTaskResult] = createWizardModifyTaskQueryMock(
@@ -357,6 +373,397 @@ describe('useRunModifyTask tests', () => {
 
     const taskModified = await screen.getByTestId('modify-task');
     expect(taskModified).toHaveTextContent('Task modified');
+    expect(screen.queryByTestId('error')).not.toBeInTheDocument();
+  });
+});
+
+const RunQuickTaskComponent = ({alertEmail, autoStart}) => {
+  const [runQuickTask] = useRunQuickTask();
+  const [reportId, setReportId] = useState();
+  const [error, setError] = useState();
+
+  const handleRunQuickTask = () =>
+    runQuickTask(
+      {
+        taskName: 'New Quick Task',
+        alertEmail,
+        autoStart,
+        configId: '08642',
+        esxiCredential: '11111',
+        smbCredential: '',
+        sshCredential: '',
+        targetHosts: '127.0.0.1, 192.168.0.1',
+        sshPort: 22,
+        startDate: mockScheduleDate,
+        startTimezone: mockScheduleTimezone,
+      },
+      mockCreationDate,
+    )
+      .then(id => {
+        if (hasValue(id)) {
+          setReportId(id);
+        } else {
+          setReportId('null');
+        }
+      })
+      .catch(setError);
+
+  return (
+    <div>
+      {reportId && (
+        <span data-testid="started-task">{`Task started with report id ${reportId}`}</span>
+      )}
+      {error && (
+        <span data-testid="error">{`There was an error in the request: ${error}`}</span>
+      )}
+      <button data-testid="wizard" onClick={handleRunQuickTask} />
+    </div>
+  );
+};
+
+describe('useRunQuickTask tests', () => {
+  test('Should create target, alert, task (but not schedule) and start task after user interaction', async () => {
+    const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+      mockScheduleDate,
+      mockScheduleTimezone,
+    );
+    const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [targetMock, targetResult] = createAdvancedWizardTargetQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [
+      createTaskMock,
+      createTaskResult,
+    ] = createAdvancedWizardCreateTaskQueryMock(
+      'New Quick Task',
+      undefined,
+      '23456',
+    );
+    const [startTaskMock, startTaskResult] = createWizardStartTaskQueryMock();
+
+    const {render} = rendererWith({
+      queryMocks: [
+        startTaskMock,
+        targetMock,
+        scheduleMock,
+        alertMock,
+        createTaskMock,
+      ],
+    });
+
+    render(
+      <RunQuickTaskComponent alertEmail={'foo@bar.com'} autoStart={'2'} />,
+    );
+
+    const button = screen.getByTestId('wizard');
+    fireEvent.click(button);
+
+    await wait();
+
+    expect(alertResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(scheduleResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(targetResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(createTaskResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(startTaskResult).toHaveBeenCalled();
+
+    const taskElement = await screen.getByTestId('started-task');
+    expect(taskElement).toHaveTextContent('Task started with report id 13245');
+    expect(screen.queryByTestId('error')).not.toBeInTheDocument();
+  });
+
+  test('Should gracefully catch error in promise chain', async () => {
+    const error = new GraphQLError('Oops. Something went wrong :(');
+
+    const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+      mockScheduleDate,
+      mockScheduleTimezone,
+    );
+    const [
+      alertMock,
+      alertResult,
+    ] = createWizardAlertQueryMock('New Quick Task', mockCreationDate, [error]);
+    const [targetMock, targetResult] = createAdvancedWizardTargetQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [
+      createTaskMock,
+      createTaskResult,
+    ] = createAdvancedWizardCreateTaskQueryMock(
+      'New Quick Task',
+      undefined,
+      '23456',
+    );
+    const [startTaskMock, startTaskResult] = createWizardStartTaskQueryMock();
+
+    const {render} = rendererWith({
+      queryMocks: [
+        startTaskMock,
+        targetMock,
+        scheduleMock,
+        alertMock,
+        createTaskMock,
+      ],
+    });
+
+    render(
+      <RunQuickTaskComponent alertEmail={'foo@bar.com'} autoStart={'2'} />,
+    );
+
+    const button = screen.getByTestId('wizard');
+    fireEvent.click(button);
+
+    await wait();
+
+    expect(alertResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(scheduleResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(targetResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(createTaskResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(startTaskResult).not.toHaveBeenCalled();
+
+    expect(screen.queryByTestId('started-task')).not.toBeInTheDocument();
+
+    const gqlError = screen.queryByTestId('error');
+
+    expect(gqlError).toHaveTextContent(
+      'There was an error in the request: Error: Oops. Something went wrong :(',
+    );
+  });
+
+  test('Should not create a schedule or start task if autoStart is 0', async () => {
+    const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+      mockScheduleDate,
+      mockScheduleTimezone,
+    );
+    const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [targetMock, targetResult] = createAdvancedWizardTargetQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [
+      createTaskMock,
+      createTaskResult,
+    ] = createAdvancedWizardCreateTaskQueryMock(
+      'New Quick Task',
+      undefined,
+      '23456',
+    );
+    const [startTaskMock, startTaskResult] = createWizardStartTaskQueryMock();
+
+    const {render} = rendererWith({
+      queryMocks: [
+        startTaskMock,
+        targetMock,
+        scheduleMock,
+        alertMock,
+        createTaskMock,
+      ],
+    });
+
+    render(
+      <RunQuickTaskComponent alertEmail={'foo@bar.com'} autoStart={'0'} />,
+    );
+
+    const button = screen.getByTestId('wizard');
+    fireEvent.click(button);
+
+    await wait();
+
+    expect(alertResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(scheduleResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(targetResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(createTaskResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(startTaskResult).not.toHaveBeenCalled();
+
+    const startTaskReportId = await screen.getByTestId('started-task');
+    expect(startTaskReportId).toHaveTextContent(
+      'Task started with report id null',
+    ); // task is not started
+    expect(screen.queryByTestId('error')).not.toBeInTheDocument();
+  });
+
+  test('Should create a schedule or but not start task if autoStart is 1', async () => {
+    const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+      mockScheduleDate,
+      mockScheduleTimezone,
+    );
+    const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [targetMock, targetResult] = createAdvancedWizardTargetQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [
+      createTaskMock,
+      createTaskResult,
+    ] = createAdvancedWizardCreateTaskQueryMock(
+      'New Quick Task',
+      '12345',
+      '23456',
+    );
+    const [startTaskMock, startTaskResult] = createWizardStartTaskQueryMock();
+
+    const {render} = rendererWith({
+      queryMocks: [
+        startTaskMock,
+        targetMock,
+        scheduleMock,
+        alertMock,
+        createTaskMock,
+      ],
+    });
+
+    render(
+      <RunQuickTaskComponent alertEmail={'foo@bar.com'} autoStart={'1'} />,
+    );
+
+    const button = screen.getByTestId('wizard');
+    fireEvent.click(button);
+
+    await wait();
+
+    expect(alertResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(scheduleResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(targetResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(createTaskResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(startTaskResult).not.toHaveBeenCalled();
+
+    const startTaskReportId = await screen.getByTestId('started-task');
+    expect(startTaskReportId).toHaveTextContent(
+      'Task started with report id null',
+    ); // task is not started
+    expect(screen.queryByTestId('error')).not.toBeInTheDocument();
+  });
+
+  test('Should not create an alert if alertEmail is empty string', async () => {
+    const [scheduleMock, scheduleResult] = createWizardScheduleQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+      mockScheduleDate,
+      mockScheduleTimezone,
+    );
+    const [alertMock, alertResult] = createWizardAlertQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [targetMock, targetResult] = createAdvancedWizardTargetQueryMock(
+      'New Quick Task',
+      mockCreationDate,
+    );
+    const [
+      createTaskMock,
+      createTaskResult,
+    ] = createAdvancedWizardCreateTaskQueryMock(
+      'New Quick Task',
+      undefined,
+      undefined,
+    );
+    const [startTaskMock, startTaskResult] = createWizardStartTaskQueryMock();
+
+    const {render} = rendererWith({
+      queryMocks: [
+        startTaskMock,
+        targetMock,
+        scheduleMock,
+        alertMock,
+        createTaskMock,
+      ],
+    });
+
+    render(<RunQuickTaskComponent alertEmail={''} autoStart={'0'} />);
+
+    const button = screen.getByTestId('wizard');
+    fireEvent.click(button);
+
+    await wait();
+
+    expect(alertResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(scheduleResult).not.toHaveBeenCalled();
+
+    await wait();
+
+    expect(targetResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(createTaskResult).toHaveBeenCalled();
+
+    await wait();
+
+    expect(startTaskResult).not.toHaveBeenCalled();
+
+    const startTaskReportId = await screen.getByTestId('started-task');
+    expect(startTaskReportId).toHaveTextContent(
+      'Task started with report id null',
+    ); // task is not started
     expect(screen.queryByTestId('error')).not.toBeInTheDocument();
   });
 });